fix(SqlServer): Fix crash on special data types in AOT
The migration crashed when reading `hierarchyid` or `geography` columns
in a Native AOT build.

The fix is to convert these types to strings on the server-side by
modifying the `SELECT` query to use `.ToString()`. The application now
only receives standard strings, avoiding the error and removing the need
for the `Microsoft.SqlServer.Types` dependency.